Extract unified plugins
