+++ src/language-markdown/unified-plugins/html-to-jsx.js
@@ -1,0 +1,20 @@
+"use strict";
+
+const mdx = require("../mdx.js");
+const { mapAst, INLINE_NODE_WRAPPER_TYPES } = require("../utils.js");
+
+function htmlToJsx() {
+  return (ast) =>
+    mapAst(ast, (node, _index, [parent]) => {
+      if (
+        node.type !== "html" ||
+        mdx.COMMENT_REGEX.test(node.value) ||
+        INLINE_NODE_WRAPPER_TYPES.includes(parent.type)
+      ) {
+        return node;
+      }
+      return { ...node, type: "jsx" };
+    });
+}
+
+module.exports = htmlToJsx;
